pyproject.toml
[tool.poetry]
name = "repl_pygame_snake"
version = "0.1.0"
description = ""
authors = ["M4RZTI4N <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
replit-play = "^0.0.23"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
